querybuilder styling
