+++ exam_frontend/src/components/QueryBuilder/QueryBuilder.css
@@ -1,0 +1,3 @@
+.querybuilder{
+    max-width: 500px;
+}
