+++ exam_frontend/src/components/TableNavBar/DrawerViewRequest.tsx
@@ -1,0 +1,26 @@
+import { FC, useState } from 'react';
+import { Drawer } from 'antd';
+import type { DrawerProps } from 'antd';
+
+interface DrawerViewRequestProps extends DrawerProps {
+    title: string;
+    onClose: () => void; 
+}
+
+const DrawerViewRequest: FC<DrawerViewRequestProps> = ({visible, title, onClose, children }) => {
+    
+    return (
+        <>
+            <Drawer
+                title={title}
+                onClose={onClose}
+                visible={visible}
+                width={800}
+            >
+                {children}
+            </Drawer>
+        </>
+    );
+};
+
+export default DrawerViewRequest;
