+++ exam_frontend/src/components/TableNavBar/TableNavbar.module.css
@@ -21,6 +21,16 @@
   color: white;
   background-color: #007bff;
 }
+.button1 {
+  width: 100px;
+  height: 50px;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 1rem;
+  color: white;
+  background-color: #007bff;
+}
 
 .button:hover {
   background-color: #0056b3;
