+++ exam_frontend/src/pages/index.tsx
@@ -7,28 +7,12 @@ import DrawerViewRequest from "./RequestPage/DrawerViewRequest";
 import { Button } from "flowbite-react";
 
 const DashboardPage: FC = function () {
-  const [visible, setVisible] = useState(false);
-
-  const showQueryBuilder = () => {
-    setVisible(true);
-  };
-
-  const closeQueryBuilder = () => {
-    setVisible(false);
-  };
- 
+  
   return (
     <div className="bg-white">
       <Statistics />
       <DasboardAssetHandler />
-      <Button onClick={showQueryBuilder} className="bg-blue-500 p-2 ml-10 h-50 w-50 font-display">QueryBuilder</Button>
-      <DrawerViewRequest
-        title="Assign"
-        onClose={closeQueryBuilder}
-        visible={visible}
-      >
-        <QueryBuilderComponent />
-      </DrawerViewRequest>
+      
     </div>
   );
 };
